First working version
I changed to specifications to make it easier implementable: The search
cannot cross buffer boundaries any more. This is OK because the primary
field of application will to be the search for disk sectors which
contain the search data, such as boot sectors, filesystem superblocks
and the like. Those never span sector boundaries.